requirements.txt
google-generativeai
tk
speechrecognition
pyaudio
pyttsx3
